feat(voucher): improve voucher management by handling exceptions and updating method signatures

+++ src/main/java/com/domus/net/domain/repository/VoucherRepository.java
@@ -13,5 +13,5 @@ public interface VoucherRepository extends baseRepository<Voucher> {
 	List<Voucher> getAllByTypeVoucher(TypeVoucher typeVoucher);
 	boolean existNumReference(String numRef);
 	void processVoucher(@Param("voucherId") Long voucherId);
-	String revertAccountsReceivable(Integer voucherId);
+	String revertAccountsReceivable(Long voucherId);
 }
